more tests for downloader
removed printing of available datasets for error handling, problems with
printing backslashes (and consequently newlines) for string to Error.
https://stackoverflow.com/questions/44780357/how-to-use-newline-n-in-f-string-to-format-output-in-python-3-6
So changed error to suggest running show_available_datasets manually
instead of running it as before. Reason for this is requesting a
non-existing dataset name (or just nothing) gave cryptic error messages,
now it is clear what went wrong. Can revisit this to bring back the
nicety of printing available datasets at a later point.