added tsne test, added to cell_attributes.csv
took the same approach for tsne as the existing approach for umap. We
should probably talk about why the former approach does not work in CI,
however, it would be a good idea to check the numbers and not just the
shape.

I added tSNE embeddings to the cell_attributes file, some difference in
especially indices (no skipped ones, as was the case in the former
version of the file), but no tests failed as a result of the change.